Refactors energypandas into its own package
